apply, update scripts

+++ update.sh
@@ -1,8 +1,14 @@
-cp ~/.zshrc ./.zshrc
-cp -r ~/.oh-my-zsh.local/ ./.oh-my-zsh.local/
-cp ~/.vimrc ./.vimrc
-cp ~/.Xdefaults ./.Xdefaults
-cp ~/.gitconfig ./.gitconfig
-cp ~/.ssh/config ./ssh-config
-cp -r ~/.config/nvim/* ./nvim/
-cp -r ~/.config/ghostty/* ./ghostty/
+#!/bin/bash
+# Update dotfiles from home directory
+
+# Copy if file exists
+[ -f ~/.zshrc ] && cp ~/.zshrc ./.zshrc
+[ -d ~/.oh-my-zsh.local ] && cp -r ~/.oh-my-zsh.local/* ./.oh-my-zsh.local/
+[ -f ~/.vimrc ] && cp ~/.vimrc ./.vimrc
+[ -f ~/.Xdefaults ] && cp ~/.Xdefaults ./.Xdefaults
+[ -f ~/.gitconfig ] && cp ~/.gitconfig ./.gitconfig
+[ -f ~/.ssh/config ] && cp ~/.ssh/config ./ssh-config
+[ -d ~/.config/nvim ] && cp -r ~/.config/nvim/* ./nvim/
+[ -d ~/.config/ghostty ] && cp -r ~/.config/ghostty/* ./ghostty/
+
+echo "Dotfiles updated from home directory"
